requirements.txt
discord.py-self
inotify